Move Paper wrapper to be consistent with component it is styling
