+++ src/components/OpenSeadragonViewer.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Paper from '@material-ui/core/Paper';
 import isEqual from 'lodash/isEqual';
 import OpenSeadragon from 'openseadragon';
 import ns from '../config/css-ns';
@@ -250,7 +249,7 @@ export class OpenSeadragonViewer extends Component {
    */
   render() {
     const {
-      windowId, children, classes, label, t,
+      windowId, children, label, t,
     } = this.props;
 
     const enhancedChildren = React.Children.map(children, child => (
@@ -270,9 +269,7 @@ export class OpenSeadragonViewer extends Component {
           ref={this.ref}
           aria-label={t('item', { label })}
         >
-          <Paper square className={classes.controls} elevation={0}>
-            { enhancedChildren }
-          </Paper>
+          { enhancedChildren }
         </section>
       </>
     );
@@ -281,7 +278,6 @@ export class OpenSeadragonViewer extends Component {
 
 OpenSeadragonViewer.defaultProps = {
   children: null,
-  classes: {},
   highlightedAnnotations: [],
   label: null,
   selectedAnnotations: [],
@@ -293,7 +289,6 @@ OpenSeadragonViewer.defaultProps = {
 OpenSeadragonViewer.propTypes = {
   canvasWorld: PropTypes.instanceOf(CanvasWorld).isRequired,
   children: PropTypes.node,
-  classes: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   highlightedAnnotations: PropTypes.arrayOf(PropTypes.object),
   label: PropTypes.string,
   selectedAnnotations: PropTypes.arrayOf(PropTypes.object),
