+++ src/containers/OpenSeadragonViewer.js
@@ -1,8 +1,6 @@
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { withTranslation } from 'react-i18next';
-import { withStyles } from '@material-ui/core';
-import { fade } from '@material-ui/core/styles/colorManipulator';
 import { withPlugins } from '../extend';
 import { OpenSeadragonViewer } from '../components/OpenSeadragonViewer';
 import * as actions from '../state/actions';
@@ -37,23 +35,7 @@ const mapDispatchToProps = {
   updateViewport: actions.updateViewport,
 };
 
-/**
- *
- * @param theme
- * @returns {{windowSideBarHeading: *}}
- */
-const styles = theme => ({
-  controls: {
-    backgroundColor: fade(theme.palette.background.paper, 0.5),
-    bottom: 0,
-    position: 'absolute',
-    width: '100%',
-    zIndex: 50,
-  },
-});
-
 const enhance = compose(
-  withStyles(styles),
   withTranslation(),
   connect(mapStateToProps, mapDispatchToProps),
   withPlugins('OpenSeadragonViewer'),
