+++ src/containers/WindowCanvasNavigationControls.js
@@ -1,6 +1,8 @@
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { withSize } from 'react-sizeme';
+import { withStyles } from '@material-ui/core';
+import { fade } from '@material-ui/core/styles/colorManipulator';
 import { withPlugins } from '../extend';
 import {
   getCanvasLabel,
@@ -18,8 +20,23 @@ const mapStateToProps = (state, { windowId }) => ({
   window: getWindow(state, { windowId }),
 });
 
+/**
+ *
+ * @param theme
+ */
+const styles = theme => ({
+  controls: {
+    backgroundColor: fade(theme.palette.background.paper, 0.5),
+    bottom: 0,
+    position: 'absolute',
+    width: '100%',
+    zIndex: 50,
+  },
+});
+
 const enhance = compose(
   connect(mapStateToProps),
+  withStyles(styles),
   withSize(),
   withPlugins('WindowCanvasNavigationControls'),
 );
